Add files via upload
This Python program is a to-do list manager that lets you add, remove, view, and mark tasks as completed. It has a user-friendly text-based interface and can be run on a Mac or online Python compilers. It's a handy tool for task organization
This is task-7 completed which is given by techno hacks